+++ crates/bin/src/main.rs
@@ -79,6 +79,7 @@ fn actual_main() -> anyhow::Result<()> {
             file_mode: args.file_mode.into(),
             read_only_prefix: args.read_only_prefix,
             once_connection: args.once_connection,
+            lessen_conflicts: args.lessen_conflicts,
         },
     )?;